refactor: extract deburrWithMap into its own module with unit tests
Move the diacritics-stripping/index-map helper out of the extension into a
dedicated module so it can be unit-tested in isolation without pulling in the
editor's React/ProseMirror component tree. Add tests covering ASCII, diacritics,
Hangul (NFD-expanded) text, ASCII matches surrounded by CJK, and surrogate pairs.
